backend/requirements.txt
anyio==4.4.0
async-property==0.2.2
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
deprecation==2.1.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.0
pycparser==2.22
python-keycloak==4.1.0
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3